Внедрил каталог автомобилей к фильтру
